user/requirements.txt
alembic
blinker
boto3
botocore
certifi
charset-normalizer
click
colorama
Flask
flask-cors
Flask-Migrate
Flask-SQLAlchemy
greenlet
idna
itsdangerous
Jinja2
jmespath
Mako
MarkupSafe
PyJWT
PyMySQL
python-dateutil
python-dotenv
requests
s3transfer
six
SQLAlchemy
typing_extensions
urllib3
Werkzeug
cryptography